chore: update LogEvent structure
